fix(views): apply preview theme's backgroundColor
This commit removes and unfortunate change from 6fad4f05fc (my change)
then had the effect of overriding the theme's background color.
The assumption was that vscode theme values can be directly reflected
but there is an option `preview.theme` where the user can change that.
Meaning a theme dark in vscode must not mean that the preview displays a
dark theme. Speaking to the intention of the change, it was about moving
styles away from general places like in `etc.ts` to the places the are
actually targeted. This would make styles more discoverable.