fix/incompatiblities with new sidebar impl (#3542)
* fix(publish): make `TreeMenu` compatible with older version of itself

This commit marks `notesLabelById` property optional so that users still
using an older version of the cli, which does not output this property,
don't get an error when running the nextjs-template.

* enhance(publish): add validation layer

This adds to begin with validation for `tree.json` data.

* chore(publis): move validation to an "fail ealier" location

This commit moves validation into `next.config.js` execution. It makes
it happens earlier and less depend on other things.

* chore(publish): improve ergonomics of handling env variables

* chore: fix typescript error

* chore(publish): add `zod` dependency to `nextjs-template`

* chore(publish): correct typo

* fix(publish): make path lookup relative to itself

* chore: use same version of `zod` across packages

* chore: pin `zod` to version `3.19.1`