fix(publish): display headings pleasantly when containing non-textonly content types  (#3525)
* fix(publish, headings): remove flex container from headings

A flex container results in its children becoming block elements.
This has many unwanted side-effects. Issues that arise after this change
are more easily fixable that issues that arise with headings being flex
containers.

* fix(publish): align heading anchors

This changes the position of the heading anchors from the left to the
right. This makes the alignment code much simpler and less fragil. From
the users POV having the anchor in the right should not effect UX. The
same approach is taking by others: https://nextra.vercel.app/features/mdx#blockquote.

* test(publish): add visual test cases for heading anchors

* chore(ci): use local user instead of root in docker container

The docker container mentioned is used for generating locally visual playwright tests.
Running the container with the local user the folders/files result not
in a surpricing permission issues.

* chore(test): update snapshots

* chore(test): adjust test case to accommodate for anchor-heading chages

Co-authored-by: Samuel Hobl <nam@hobl.at>