chore(publish): disable e2e test with NEXT_PUBLIC_ASSET_PREFIX set
The issue is that add that env var results into css styles not being
rendered and therefore changing the form of that result that we want to
test with visual regression testing. Maybe we need to seperate VRT from
e2e testing which appear to me to be actually not the same. VRT does not
require interaction with the page.