enhance/nextjs playwright tests (#3580)
* chore: remove unused `cypress.json` file

* chore: rename playwright test folder to `e2e`

* chore(publish): add `NODE_ENV` to `env` handler

* chore(publish, test): correct ts types

* chore(publish): split env into client and schema

* chore(publish): test to make sure NEXT_PUBLIC_ASSET_PREFIX env is used

* chore(publish): correct test description

* chore(publish): update playwright to use new `toHaveScreenshot` api

This is a more ergonomic API and want to use that inside [[dendron://dendron.docs/pkg.nextjs-template.qa.test]].

* chore(publish): allow passing arguments to `cli:test:template:docker` command

This commit aims to allow passing arguments to the mentions command like
so: `yarn cli:test:template:docker --update-snapshots`.

It appears docker does not support positional arguments therefore prepending `$0` so that the first argument gets applied.

* chore(publish): disable e2e test with NEXT_PUBLIC_ASSET_PREFIX set

The issue is that add that env var results into css styles not being
rendered and therefore changing the form of that result that we want to
test with visual regression testing. Maybe we need to seperate VRT from
e2e testing which appear to me to be actually not the same. VRT does not
require interaction with the page.

* chore(publish): revert changes

Lets keep this how it was for now.

* chore(publish): update visual snapshots

* chore(publish): revert logic to allow NODE_ENV=test

* chore(publish): consider `NEXT_PUBLIC_ASSET_PREFIX` when running e2e test

* chore: adjust command to forward arguments intuitively

This commit aims to allow `yarn run ci:test:template:docker -u` or any
other multiple arguments appendix.

* chore(publish): update e2e VRT snapshots

The cause of the changes is an update from playwright and its browser
dependencies.

* chore(publish): add e2e test for `NEXT_PUBLIC_ASSET_PREFIX`

* chore(publish): fix react key warning

* chore(publish, e2e): rephrase test case description