fix: window path slash (#3755)
* fix: window path slash

* make `normalizeUnixPath` the main function for its kind of purpose

* embed slash library into our utils

The rational is that the latest version of `slash` is build as a pure ES
module which we currently don't support as a dependency.

* update comment

* replace `slash` with `normalize-path`

To fit the purpose of `normalizeUnixPath` this commit switches to `normalize-path` which not only has the slash conversion but also path simplification and matches `path.posix.normalize` much closer.