fix(views): calendar view broken after theme toggle (#3765)
* fix(views): broken calendar view on theme toggle

This commit just removes a default value for the `initialTheme`
parameter inside the views util `getWebviewContent`.

`genVSCodeHTMLIndex`, which is called with a forwarded `initialTheme`,
has a codepath for an undefined `initialTheme` parameter.
It means that no theme is going to override the default theme and
therefore a certain ordering of css file references is taking place.

So by always providing a default theme this codepath is never taken and
results into a broken calendar view when toggling theme value.

Removing that makes sure that the default theme is setup in way that
keeps proper style ordering in place.

That default theme was added as a solution for
[[dendron://private/task.2022.05.31.fix-styling-for-plugin-views]] does
not seam to be necessary for that.

* chore(internal): fix invalid test-workspace note