fix(publish): display full breadcrumb path for sidebar excluded notes (#3742)
* fix(publish): displaying breadcrumb on sidebar hidden notes

This commit aim's to make breadcumbs be shown when the note is hidden from the sidebar.
This can happen with note properties like `has_collection` or `nav_exclude_children`.

This is solved by moving that logic out of sidebar generator into
tree-utils and the directly effected component (`DendronTreeMenu`).

* chore: update test cases

* chore(internal, e2e): expand visual regression tests for breadcrumb display

* chore: revert `264b2e35c`

The plan is the approach the issue from a different strategy.

* Revert "chore: update test cases"

This reverts commit 6fbf55a29e7e0e7cc3cb370bbe7ec766ef46b94a.

* fix(publish): displaying breadcrumb on sidebar hidden notes